requirements.txt
selene == 2.0.0rc9
pytest == 8.2.2
allure-pytest == 2.13.5
requests == 2.32.3